+++ source/DSCResources/DSC_SqlAGReplica/DSC_SqlAGReplica.psm1
@@ -765,4 +765,3 @@ function Test-TargetResource
     return $result
 }
 
-Export-ModuleMember -Function *-TargetResource