+++ source/DSCResources/DSC_SqlAgentOperator/DSC_SqlAgentOperator.psm1
@@ -335,4 +335,3 @@ function Test-TargetResource
     $isOperatorInDesiredState
 }
 
-Export-ModuleMember -Function *-TargetResource