+++ source/DSCResources/DSC_SqlAlwaysOnService/DSC_SqlAlwaysOnService.psm1
@@ -268,4 +268,3 @@ function Test-TargetResource
     return $isInDesiredState
 }
 
-Export-ModuleMember -Function *-TargetResource