+++ source/DSCResources/DSC_SqlLogin/DSC_SqlLogin.psm1
@@ -825,4 +825,3 @@ function Set-SQLServerLoginPassword
     }
 }
 
-Export-ModuleMember -Function *-TargetResource