+++ source/DSCResources/DSC_SqlReplication/DSC_SqlReplication.psm1
@@ -891,4 +891,3 @@ function Get-SqlLocalServerName
     }
 }
 
-Export-ModuleMember -Function *-TargetResource