+++ source/DSCResources/DSC_SqlScript/DSC_SqlScript.psm1
@@ -385,4 +385,3 @@ function Test-TargetResource
     }
 }
 
-Export-ModuleMember -Function *-TargetResource