+++ CHANGELOG.md
@@ -10,12 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SqlServerDsc
   - The module manifest property `DscResourcesToExport` now updates automatically
     using the pipeline.
+  - Removed `Export-ModuleMember` from DSC resource that still had it.
 
 ### Fixed
 
 - SqlAGDatabase
   - Fix for issue ([issue #1492](https://github.com/dsccommunity/SqlServerDsc/issues/1492))
-    added AutomaticSeeding for this resource. In Set-TargetResource added logic that looks 
+    added AutomaticSeeding for this resource. In Set-TargetResource added logic that looks
     at all replicas of an availability group. When automatic seedig is found, it will use that.
   - Lots of extra tests to check AutomaticSeeding.
   - The parameter `BackupPath` is still needed just in case a database never has been backuped before.